+++ src/main/java/com/app/camvil/dto/requestdto/CommentCreateRequestDTO.java
@@ -1,29 +1,29 @@
 package com.app.camvil.dto.requestdto;
 
 public class CommentCreateRequestDTO {
-    private int userId;
-    private int boardId;
+    private long userId;
+    private long boardId;
     private String commentContent;
 
-    public CommentCreateRequestDTO(int userId, int boardId, String commentContent) {
+    public CommentCreateRequestDTO(long userId, long boardId, String commentContent) {
         this.userId = userId;
         this.boardId = boardId;
         this.commentContent = commentContent;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
 
-    public int getBoardId() {
+    public long getBoardId() {
         return boardId;
     }
 
-    public void setBoardId(int boardId) {
+    public void setBoardId(long boardId) {
         this.boardId = boardId;
     }
